# ms-movements/src/main/java/dev/guillermosg/msmovements/domain/model/UpdateStatusAccountRequest.java
package dev.guillermosg.msmovements.domain.model;

import lombok.*;

import java.io.Serializable;

/**
 * UpdateStatusAccountRequest model
 */

@Getter
@Setter
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UpdateStatusAccountRequest implements Serializable {

    private static final long serialVersionUID = 1L;

    private Boolean estado;
    private Integer cuentaId;
}